LPS-99657 Fix typos
