+++ modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/xstream/configurator/SegmentsXStreamConfigurator.java
@@ -12,13 +12,14 @@
  * details.
  */
 
-package com.liferay.asset.display.page.internal.xtream.configurator;
+package com.liferay.segments.internal.xstream.configurator;
 
-import com.liferay.asset.display.page.model.AssetDisplayPageEntry;
 import com.liferay.exportimport.kernel.xstream.XStreamAlias;
 import com.liferay.exportimport.kernel.xstream.XStreamConverter;
 import com.liferay.exportimport.kernel.xstream.XStreamType;
 import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.segments.model.impl.SegmentsEntryImpl;
+import com.liferay.segments.model.impl.SegmentsExperienceImpl;
 import com.liferay.xstream.configurator.XStreamConfigurator;
 
 import java.util.List;
@@ -27,11 +28,10 @@ import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Pavel Savinov
+ * @author Eduardo García
  */
 @Component(immediate = true, service = XStreamConfigurator.class)
-public class AssetDisplayPageXStreamConfigurator
-	implements XStreamConfigurator {
+public class SegmentsXStreamConfigurator implements XStreamConfigurator {
 
 	@Override
 	public List<XStreamType> getAllowedXStreamTypes() {
@@ -51,8 +51,8 @@ public class AssetDisplayPageXStreamConfigurator
 	@Activate
 	protected void activate() {
 		_xStreamAliases = new XStreamAlias[] {
-			new XStreamAlias(
-				AssetDisplayPageEntry.class, "AssetDisplayPageEntry")
+			new XStreamAlias(SegmentsEntryImpl.class, "SegmentsEntry"),
+			new XStreamAlias(SegmentsExperienceImpl.class, "SegmentsExperience")
 		};
 	}
